requirements.txt
bs4~=0.0.2
beautifulsoup4~=4.12.3
langchain~=0.2.6
chromadb~=0.5.3
jsonlines~=4.0.0
gradio~=4.44.0
SQLAlchemy~=2.0.31
python-dotenv~=1.0.1
openai~=1.35.10